refactor component DropdownDateRange
