+++ HealthCareApp/Components/Dropdown/DropdownDateRange.razor
@@ -6,11 +6,38 @@
             data-bs-auto-close="outside"
             aria-expanded="false">
         <i class="bi bi-calendar3 text-@IconColor.ToString()"></i>
-        @DateRange
+        @_dateRangeLabel
         <i class="bi bi-chevron-down ps-2"></i>
     </button>
-    <ul class="dropdown-menu dropdown-menu-end border-light shadow-lg"
+    <ul class="dropdown-menu dropdown-menu-end border-light shadow-lg ps-2"
         aria-labelledby="dropdownDateRange">
-        @ChildContent
+
+        <li class="d-flex pe-2 gap-2">
+            <input class="form-control form-control-sm"
+                    id="Start"
+                    type="date"
+                    placeholder="Start date"
+                    @bind-value="_dateTimeRange.Start" />
+
+            <input class="form-control form-control-sm"
+                    id="End"
+                    type="date"
+                    placeholder="End date"
+                    @bind-value="_dateTimeRange.End" />
+            <button type="submit"
+                    class="btn btn-primary btn-sm"
+                    title="Confirm"
+                    @onclick="() => ChangeDateAsync()">
+                <i class="bi bi-check-lg"></i>
+            </button>
+        </li>
+        @if (!_isValidDateRange)
+        {
+            <li class="d-flex px-2 gap-2">
+                <span class="validation-message">
+                    Invalid date range, please check!
+                </span>
+            </li>
+        }
     </ul>
 </div>
