--- HealthCareApp/Components/Dropdown/DropdownDateRangeOption.razor
@@ -1,29 +1,0 @@
-﻿<ul class="ps-2">
-    <li class="d-flex pe-2 gap-2">
-        <input class="form-control form-control-sm"
-               id="Start"
-               type="date"
-               placeholder="Start date"
-               @bind-value="DateTimeRange.Start" />
-
-        <input class="form-control form-control-sm"
-               id="End"
-               type="date"
-               placeholder="End date"
-               @bind-value="DateTimeRange.End" />
-        <button type="submit"
-                class="btn btn-primary btn-sm"
-                title="Confirm"
-                @onclick="() => ChangeDateAsync()">
-            <i class="bi bi-check-lg"></i>
-        </button>
-    </li>
-    @if (!_isValidDateRange)
-    {
-        <li class="d-flex px-2 gap-2">
-            <span class="validation-message">
-                Invalid date range, please check!
-            </span>
-        </li>
-    }
-</ul>
