+++ HealthCareApp/Pages/DashboardPage/Dashboard.razor
@@ -23,8 +23,8 @@
                             <div class="col-sm-12 col-md-6 col-lg-4 d-flex flex-column">
                                 <ActiveUsers />
                             </div>
-                            <div class="col-sm-12 col-md-6 col-lg-4 d-flex">
-                                <ChartMopUsage />
+                            <div class="col-sm-12 col-md-6 col-lg-4 d-flex flex-column">
+                                <ChartMopUsage DateTimeRange="@_dateTimeRange" />
                             </div>
                         </div>
                     </PageMiddleWrapper>
